WEB-70: add 'method' to Service, move to services/__init__
